Expand sample program and move it to com.example package.
The com.example package is based on the well-known domain example.com,
which is defined specifically for use in examples, and won't clash with any
other packages.

Breaks the sample main method out to a PersonEditor class, so we're not referencing
just the Form class.

Adds a PersonEditorWithSwingExplorer example launcher, to make it convenient to run
the sample under Swing Explorer from within IDEs like IntelliJ.